refactor: change page components to tsx, fix types
