+++ pages/work/dummyData.ts
@@ -1,10 +1,7 @@
-import Header from '../../components/shared/Header';
-import Container from '../../components/shared/Container';
-import Project from '../../components/shared/Project';
-import ChapterTable from '../../components/scoped/work/ChapterTable';
-import ChapterFeature from '../../components/scoped/work/ChapterFeature';
+import { ChapterType } from '../../components/scoped/work/ChapterTable';
+import { ProjectType } from '../../components/shared/Project';
 
-const exampleProjects = [
+export const exampleProjects: ProjectType[] = [
   {
     image: 'http://placekitten.com/g/1000/600',
     title: 'Website',
@@ -23,7 +20,7 @@ const exampleProjects = [
   },
 ];
 
-const exampleChapters = [
+export const exampleChapters: ChapterType[] = [
   {
     logoUrl: 'http://placekitten.com/g/100/100',
     name: 'Kitten Tech',
@@ -389,25 +386,3 @@ const exampleChapters = [
     projects: exampleProjects,
   },
 ];
-
-function Work() {
-  return (
-    <main>
-      <Header title="Our Work" />
-      <Container>
-        <h2>National Initiatives</h2>
-        {exampleProjects.map(item => (
-          <Project key={item.title} {...item} />
-        ))}
-      </Container>
-      <Container>
-        <h2>Our Chapters</h2>
-        <ChapterTable items={exampleChapters}>
-          {selectedItem => <ChapterFeature {...selectedItem} />}
-        </ChapterTable>
-      </Container>
-    </main>
-  );
-}
-
-export default Work;
